+++ metadata/src/test/java/org/apache/kafka/controller/BrokersToIsrsTest.java
@@ -19,6 +19,7 @@ package org.apache.kafka.controller;
 
 import org.apache.kafka.common.Uuid;
 import org.apache.kafka.common.utils.LogContext;
+import org.apache.kafka.server.common.TopicIdPartition;
 import org.apache.kafka.controller.BrokersToIsrs.PartitionsOnReplicaIterator;
 import org.apache.kafka.timeline.SnapshotRegistry;
 import org.junit.jupiter.api.Test;
